[frontend] move components out of pages routing
